Refactor: Babel - Split same-file static evaluation
